Separated dynamic import functions into separate file.
