+++ Makefile
@@ -10,6 +10,7 @@ s3-config:
 	aws s3 cp src/config/utils.js s3://$(BUCKET_NAME)/scripts/threejs/config/utils.js
 	aws s3 cp src/config/uiPanelConfig.js s3://$(BUCKET_NAME)/scripts/threejs/config/uiPanelConfig.js
 	aws s3 cp src/config/attachUIListeners.js s3://$(BUCKET_NAME)/scripts/threejs/config/attachUIListeners.js
+	aws s3 cp src/config/dynamicImports.js s3://$(BUCKET_NAME)/scripts/threejs/config/dynamicImports.js
 
 s3-drawing:
 	aws s3 cp src/drawing/chartConfig.js s3://$(BUCKET_NAME)/scripts/threejs/drawing/chartConfig.js
